Adding channel and velocity as explicit class members and adding their setters
At the moment the only way to set channel and velocity for a Note is through class constructor. But the constructor calls for a 'name' object including those items. I've tried to conform to this format but I'm getting error when trying to supply conforming object. So it turned out the only way to actually set the channel for a Note was by adding members and accessors for conveniency. The solution works for me in the app where I need to split Notes to different channels