Added ability to set instrument_nr and set instrument name based on GM instr no.
I didn't seem to be possible to set instrument_nr. It was hardcoded as 1 and setting it as per docs didn't work. Also the list of instrument names didn't seem to be utilized so I propose changing the constructor argument from name to instrument_nr and then the name is being auto-set based on id using the GM list. Not really sure how to handle out of range numbers here (are they allowed by midi standard ect) so just setting 'Unknown Instrument' as name